+++ index.js
@@ -1,7 +1,6 @@
 module.exports = {
   Switch: require("./Switch"),
   ColorPicker: require("./ColorPicker"),
-  TagsInput: require("./TagsInput"),
   YoutubeVideo: require("./YoutubeVideo"),
   VimeoVideo: require("./VimeoVideo"),
   LoadingSpinner: require("./LoadingSpinner"),